CollectionGatedReferenceModule uint test

+++ contracts/core/modules/reference/CollectionGateReferenceModule.sol
@@ -8,6 +8,7 @@ import {ModuleBase} from '../ModuleBase.sol';
 import {IERC721} from '@openzeppelin/contracts/token/ERC721/IERC721.sol';
 import {IERC1155} from '@openzeppelin/contracts/token/ERC1155/IERC1155.sol';
 import {EnumerableSet} from '@openzeppelin/contracts/utils/structs/EnumerableSet.sol';
+import 'hardhat/console.sol';
 
 /**
  * @title CollectionGatedReferenceModule
@@ -16,7 +17,6 @@ import {EnumerableSet} from '@openzeppelin/contracts/utils/structs/EnumerableSet
  * @notice A simple reference module that validates that comments or mirrors originate is the owner of a given collection.
  */
 contract CollectionGatedReferenceModule is ModuleBase, IReferenceModule {
-
     constructor(address hub) ModuleBase(hub) {}
 
     /**
@@ -27,8 +27,6 @@ contract CollectionGatedReferenceModule is ModuleBase, IReferenceModule {
         uint256 pubId,
         bytes calldata data
     ) external override returns (bytes memory) {
-        // decode
-
         return new bytes(0);
     }
 
@@ -46,12 +44,7 @@ contract CollectionGatedReferenceModule is ModuleBase, IReferenceModule {
         address commentCreator = IERC721(HUB).ownerOf(profileId);
         (address collection, uint256[] memory tokenIdSet) = abi.decode(data, (address, uint256[]));
         if (collection == address(0)) revert Errors.InitParamsInvalid();
-
-        _checkCollectionValidity(
-            commentCreator,
-            collection,
-            tokenIdSet
-        );
+        _checkCollectionValidity(commentCreator, collection, tokenIdSet);
     }
 
     /**
@@ -68,18 +61,14 @@ contract CollectionGatedReferenceModule is ModuleBase, IReferenceModule {
         address mirrorCreator = IERC721(HUB).ownerOf(profileId);
         (address collection, uint256[] memory tokenIdSet) = abi.decode(data, (address, uint256[]));
         if (collection == address(0)) revert Errors.InitParamsInvalid();
-        _checkCollectionValidity(
-            mirrorCreator,
-            collection,
-            tokenIdSet
-        );
+        _checkCollectionValidity(mirrorCreator, collection, tokenIdSet);
     }
 
     function _checkCollectionValidity(
         address user,
         address collection,
-        uint[] memory tokenIdSet
-    ) internal view{
+        uint256[] memory tokenIdSet
+    ) internal view {
         bool isErc721 = IERC721(collection).supportsInterface(type(IERC721).interfaceId);
         if (!isErc721) {
             if (!IERC1155(collection).supportsInterface(type(IERC1155).interfaceId))
